setup: Clean up asm functions
print_char() is only called from C, and there is very little ASM code where it
can be used, given how little is 64bit.  Reimplement print_char() in C, where
the optimiser does a rather better job.

stgi() isn't used at all, and would be better implemented with a static inline
as it is shorter than the call instruction required to invoke it.

Add __noreturn and unreachable() definitions, which help the compiler to
discard unnecessary epilogue code.  Annotate die() with unreachable(), and
setup() and lz_exit() with __noreturn.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>