entry: Streamline asm
Alter the EFER_* and VM_CR_* bits to be masks rather than bit positions.  This
is more common, and is how all of them are used.

Fold adjacent andl's over the same register.  Don't bother loading %fs/%gs; C
doesn't use them, and the only reason that inline assembly might is if they
are different to %ds.

Don't bother enabling PGE in CR4 as it isn't used, and allows for imm8
encoding of the or instruction.  Make use of the high byte registers when it
allows imm32 to reduce to imm8.

Don't bother enabling CR0.ET as it is hardwired, but set TS to catch any
floating-point or vector instruction instead of silently corrupting existing
state.  There is also no need to jump forward after the write to CR0, as it is
immediately followed by another jump.

For the lz_exit path, all values are actually 32bit.  Only undo the control
register modifications we made, which simplifies things substantially.

  add/remove: 0/0 grow/shrink: 0/2 up/down: 0/-91 (-91)
  Function                                     old     new   delta
  _entry                                       149     133     -16
  lz_exit                                      129      54     -75
  Total: Before=49309, After=49218, chg -0.18%

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>