entry: Fold GDTR into GDT
The GDT is only loaded in 32bit mode.  The 64 suffixes are misleading, and the
lgdt instruction will only take a .long base, rather than .quad.  This also
means GDTR can be folded into the NUL segment which is a common 32bit trick.

Rename these to gdt/gdtr respectively (to drop the misleading suffixes), drop
the unused gdt_desc64_end symbol, and use a local symbol for gdt_end to make
the disassembly easier to follow.

This drops 32 bytes of data.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Krystian Hebel <krystian.hebel@3mdeb.com>
Tested-by: Krystian Hebel <krystian.hebel@3mdeb.com>