entry: Reduce data segment loading
Using a 32bit register to load segment selectors is shorter to encode
and faster to execute.  There is no need to reload the segments after
the jump between 32 and 64bit mode, so drop those blocks entirely.

Furthermore, as the GDT layout matches the SKINIT initial state, there
is no need to reload %ss.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Krystian Hebel <krystian.hebel@3mdeb.com>
Tested-by: Krystian Hebel <krystian.hebel@3mdeb.com>