entry: Rearrange the GDT to match AMD's spec
c/s 1e6fac00b rearranged the GDT yet again, and in the process making it no
longer compatible with the state after SKINIT.  This appears to have been for
debugging reasons.

Switch back to an SKINIT-compatible layout, which will reduce the number of
segments we need to reload.

Furthermore, set the Accessed bits.  We are not using segmented virtual
memory, so don't want to force the processor to set them on first use.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Krystian Hebel <krystian.hebel@3mdeb.com>
Tested-by: Krystian Hebel <krystian.hebel@3mdeb.com>