setup: Remove opencoded calculation of the location of lz_header
Make the lz_header symbol global, and extern it in C.  This removes the need
for the local lz_header pointer.

While at it, introduce some linkage macros and provide ELF size/type
information for lz_header, which causes objdump not to try and disassemble the
header.  Drop the unused LZ_HEADER_{SIZE,OFFSET} constants.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Krystian Hebel <krystian.hebel@3mdeb.com>
Tested-by: Krystian Hebel <krystian.hebel@3mdeb.com>