build: Link lz_header more sensibly
Alter the linker script to contain the data sections (at least a cache line
away from text), including subsections which already exist such as
.rodata.str.1.  Drop the lzdata section which is obsolete since c/s bd9c4c210.

This means that variables in setup.c no longer need inserting into .text, so
drop the __text annotation.

Build all objects with debug symbols, to make the resulting ELF more useful to
various tools, and generate the final binary with a single objcopy which
strips all extraneous sections.  Take the opportunity to drop mkbin.sh and its
various droppings.

Do not link the final binary as PIE.  Any outstanding relocations will be
fatal to the eventual running.  This means that .dynamic and .got.plt should
not be captured in the linker script.

Finally, add two sanity checks concerning the image itself.  First, inspect
lz_header.bin to check that the UUID is in the expected location.

Second, use a linker assertion to check that we don't exceed the SLB maximum
size.  For now, the size is limited by the statically allocated objects later
in the image.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Krystian Hebel <krystian.hebel@3mdeb.com>
Tested-by: Krystian Hebel <krystian.hebel@3mdeb.com>