entry: Drop lz_exit and return from setup()
Currently, asm calls setup(), and setup eventually calls lz_exit() to
re-enter asm.  This would be more simple to follow if setup() returned,
but that complicates the passing of the lz_exit() parameters.

Luckily, there is the "small object return ABI" in SYSV which allows us
to return two pointers directly in registers.  Document both sides of
this, as it is more magic that most people are familiar with.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Tested-by: Daniel P. Smith <dpsmith@apertussolutions.com>