setup: Don't pass lz_base pointers everywhere
Currently, a pointer to the base of LZ is passed from asm into setup(),
then stashed in a global C variable, and then passed back into asm as a
parameter to lz_exit().  This is all quite inefficient.

Make sl_header global, and expose it to C.  This is by definition the
base of LZ, although retaining the lz_base name is probably easiest from
the programmers point of view.

Strip all lz_base parameter passing.  As %ebp is caller-saved, is still
correct at the beginning of lz_exit, and no further restoration is
required.

Introduce a macro to make lz_base usable in C as before, but in terms of
sl_header.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Tested-by: Daniel P. Smith <dpsmith@apertussolutions.com>