early_tpm: Fix big-endian helpers
First of all, __get_unaligned_be32() has undefined behaviour as "p[0] << 24"
causes p[0] to be promoted to int rather than unsigned int, resulting with a
shift into the sign bit.

Secondly, taking val by pointer forces it to be spilled to memory, which is
why __get_unaligned_be32() was out-of-lined despite recognising that the
operation can be implemented using the bswap instruction.

Delete be_byteshift.h, and implement the transformations locally in terms of
scalar values.

GCC now takes the opportunity to implement the 16 bit swaps using `rolw $8,
mem16` in place in the original buffer, and to compute cpu_to_be32(const) at
compile time, rather than spilling a constant to the stack, and calling
__get_unaligned_be32() to get the answer.

  add/remove: 0/1 grow/shrink: 0/3 up/down: 0/-165 (-165)
  Function                                     old     new   delta
  __get_unaligned_be32                           5       -      -5
  tis_recv                                     287     258     -29
  tpm2_null_auth                                52      22     -30
  tpm2_extend_pcr                              489     388    -101
  Total: Before=48606, After=48441, chg -0.34%

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Reviewed-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Tested-by: Daniel P. Smith <dpsmith@apertussolutions.com>