build: Introduce a headers section
Currently, the link order of object matters, because head.S's .text section
has to end up first.  Instead, put the headers in a dedicated section and use
the linker file to ensure that they end up first.

This removes the need for head.o to be the first object the linker sees.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>