Helpers for integer/pointer conversion
For 32 and 64-bit clean code, pointers an integers should be cast through
uintptr_t, and specifically not u32 or u64.  When manipulating cast pointers
as integers, make sure to use UL and not ULL to avoid promotion to long long
in 32bit.

Introduce the _p() and _u() helpers, and update code to use them, dropping
unnecessary casts as appropriate.

Switch mmio_base_addr to be uint32_t - the pointer is under the 4G
boundary (as we don't take #PF when trying to use it) and this will be a hard
limit for 32bit.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>