iommu: Implementation of early IOMMU
IOMMU is enabled in separate steps:
* IOMMU is enabled, but fails to read its commands from the command
  buffer, as it is still locked by SLB,
* IOMMU enters fail-safe state in which all device accesses are blocked,
* DEV SLB protection is disabled,
* IOMMU is re-enabled, this time it can read data previously protected
  by SLB.

Devices can access RAM (except for SLB) before IOMMU is initially enabled,
but not after that.

This implementation uses 8 kB Device Table. 4 kB table did not cover all
of the bus 0 devices. All devices that are not covered have master-aborted
memory accesses (which is good), as well as interrupts (also good, most of
the times). One of such devices was LPC, under which PIT is implemented.
As a result, legacy timer interrupts were not received by the CPU, which
resulted in kernel panic.

Tested with legacy ISA DMA.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>