build: move bootloader data out of measured block
Bootloader must pass pointer to the zeropage, which is not constant. To
keep hash values independent of memory layout, those variables are moved
to a separate section out of SL (but still inside SLB). Length in SL
header is no longer set to maximum possible value, it points to the end
of constant data.

GRUB2 code must be changed accordingly.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>