sha256: API cleanup and unit test
struct sha256_state is internal and not used by callers.  Move it from
sha256.h to sha256.c.  The inital values are used once, so drop the defines
entirely.

For sha256sum(), use C99 array bounds syntax to help the compiler determine
that the destination buffer is large enough, and constify the input pointer as
it is not modified.

Add a unit test with some easy test vectors.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>