types: Drop rarely used types
Having a mix between uint* and u* is a pain for unit testing.  As u* is by far
the most common, use it consistently and drop the uint* aliases, and fixing up
the 3 uers.

Also drop the min/max macros.  They aren't used, and don't want to be written
like that if they do get used.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>