extend_all.sh: support kernels with embedded initramfs
Modify test for number of command line arguments. For one argument,
measure the file pointed by that argument assuming it is a Linux kernel
and simulate a PCR extend operation for just the LZ and the kernel's hash.

This commit also refactors code for extend_shaXX functions and adds more
helpers - separate functions for calculating hashes of measured parts of
LZ and kernel.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>