event_log.c: make the log format compatible with TXT
In order to simplify adding new entries and to minimize the number of
differences between AMD and Intel, the event logs have actually two headers,
both for TPM 1.2 and 2.0.

For TPM 1.2 this is TCG_PCClientSpecIDEventStruct [1] with Intel's own
TXT-specific header embedded inside its 'vendorInfo' field. The offset to
this field is added to the base address in AMD path in kernel, making the
code for adding new events the same for both vendors.

TPM 2.0 in TXT uses HEAP_EVENT_LOG_POINTER_ELEMENT2_1 structure, which is
normally constructed on the TXT stack [2]. For AMD, this structure is put
inside TCG_EfiSpecIdEvent [3], also in 'vendorInfo' field. The actual offset
to this field depends on number of hash algorithms supported by the event
log.

[1] https://www.trustedcomputinggroup.org/wp-content/uploads/TCG_PCClientImplementation_1-21_1_00.pdf
[2] http://www.intel.com/content/dam/www/public/us/en/documents/guides/intel-txt-software-development-guide.pdf
[3] https://trustedcomputinggroup.org/wp-content/uploads/TCG_PCClientSpecPlat_TPM_2p0_1p04_pub.pdf

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>