Parse bootloader data in the form of tags
This patch was created mainly to deal with compatibility issues between
bootloader and Landing Zone appearing after every feature added. It should
give the possibility to add more pieces of information passed from the
bootloader in the future without breaking booting of older versions of LZ,
starting with this one.

Also, from now on it is the bootloader's task to calculate and pass the LZ
hash(es), there are no longer precalculated and included in the file.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>